refactored test code wrt to SecurityUser to fix failing test (DonationBatchCRUDServiceTest)
