+++ test/repository/ComponentRepositoryTest.java
@@ -1,7 +1,6 @@
 package repository;
 
 import java.io.File;
-import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
@@ -9,8 +8,6 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import javax.sql.DataSource;
 
 import model.component.Component;
 import model.component.ComponentStatus;
@@ -20,37 +17,20 @@ import model.componentmovement.ComponentStatusChangeType;
 import model.componenttype.ComponentType;
 import model.donation.Donation;
 
-import org.dbunit.database.DatabaseConfig;
-import org.dbunit.database.DatabaseDataSourceConnection;
-import org.dbunit.database.IDatabaseConnection;
 import org.dbunit.dataset.IDataSet;
 import org.dbunit.dataset.xml.FlatXmlDataSetBuilder;
-import org.dbunit.ext.hsqldb.HsqldbDataTypeFactory;
-import org.dbunit.operation.DatabaseOperation;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.TestingAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.transaction.AfterTransaction;
-import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.transaction.annotation.Transactional;
 
-import security.BsisUserDetails;
+import suites.DBUnitContextDependentTestSuite;
 
 /**
  * Test using DBUnit to test the ComponentRepository
  */
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(locations = "file:**/applicationContextTest.xml")
-@Transactional
-@WebAppConfiguration
-public class ComponentRepositoryTest {
+public class ComponentRepositoryTest extends DBUnitContextDependentTestSuite {
 	
 	@Autowired
 	ComponentRepository componentRepository;
@@ -59,60 +39,12 @@ public class ComponentRepositoryTest {
 	DonationRepository donationRepository;
 	
 	@Autowired
-	UserRepository userRepository;
-	
-	@Autowired
 	ComponentStatusChangeReasonRepository componentStatusChangeReasonRepository;
 	
-	@Autowired
-	private DataSource dataSource;
-	
-	private IDataSet getDataSet() throws Exception {
+	@Override
+	protected IDataSet getDataSet() throws Exception {
 		File file = new File("test/dataset/ComponentRepositoryDataset.xml");
 		return new FlatXmlDataSetBuilder().setColumnSensing(true).build(file);
-	}
-	
-	private IDatabaseConnection getConnection() throws SQLException {
-		IDatabaseConnection connection = new DatabaseDataSourceConnection(dataSource);
-		DatabaseConfig config = connection.getConfig();
-		config.setProperty(DatabaseConfig.PROPERTY_DATATYPE_FACTORY, new HsqldbDataTypeFactory());
-		return connection;
-	}
-	
-	@Before
-	public void init() throws Exception {
-		IDatabaseConnection connection = getConnection();
-		try {
-			IDataSet dataSet = getDataSet();
-			DatabaseOperation.CLEAN_INSERT.execute(connection, dataSet);
-			initSpringSecurityUser(); // must be done after data has been inserted successfully
-		}
-		finally {
-			connection.close();
-		}
-	}
-	
-	@AfterTransaction
-	public void after() throws Exception {
-		IDatabaseConnection connection = getConnection();
-		try {
-			clearSpringSecurityUser(); // reverse the initiation of the Spring Security user
-			IDataSet dataSet = getDataSet();
-			DatabaseOperation.DELETE_ALL.execute(connection, dataSet);
-		}
-		finally {
-			connection.close();
-		}
-	}
-	
-	public void initSpringSecurityUser() throws Exception {
-		BsisUserDetails user = new BsisUserDetails(userRepository.findUserById(1l));
-		TestingAuthenticationToken auth = new TestingAuthenticationToken(user, "Credentials");
-		SecurityContextHolder.getContext().setAuthentication(auth);
-	}
-	
-	public void clearSpringSecurityUser() throws Exception {
-		SecurityContextHolder.getContext().setAuthentication(null);
 	}
 	
 	// TODO: test the following methods
