+++ test/suites/DBUnitContextDependentTestSuite.java
@@ -1,0 +1,92 @@
+package suites;
+
+import java.sql.SQLException;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.sql.DataSource;
+
+import model.user.User;
+
+import org.dbunit.database.DatabaseConfig;
+import org.dbunit.database.DatabaseDataSourceConnection;
+import org.dbunit.database.IDatabaseConnection;
+import org.dbunit.dataset.IDataSet;
+import org.dbunit.ext.hsqldb.HsqldbDataTypeFactory;
+import org.dbunit.operation.DatabaseOperation;
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.TestingAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.transaction.AfterTransaction;
+import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.transaction.annotation.Transactional;
+
+import repository.UserRepository;
+import security.BsisUserDetails;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(locations = "file:**/applicationContextTest.xml")
+@Transactional
+@WebAppConfiguration
+public abstract class DBUnitContextDependentTestSuite {
+  
+  protected static final String ADMIN_USERNAME = "admin";
+  protected User adminUser;
+  
+  @PersistenceContext
+  protected EntityManager entityManager;
+  
+  @Autowired
+  UserRepository userRepository;
+    
+  @Autowired
+  private DataSource dataSource;
+  
+  protected abstract IDataSet getDataSet() throws Exception;
+
+  private IDatabaseConnection getConnection() throws SQLException {
+    IDatabaseConnection connection = new DatabaseDataSourceConnection(dataSource);
+    DatabaseConfig config = connection.getConfig();
+    config.setProperty(DatabaseConfig.PROPERTY_DATATYPE_FACTORY, new HsqldbDataTypeFactory());
+    return connection;
+  }
+
+  @Before
+  public void init() throws Exception {
+    IDatabaseConnection connection = getConnection();
+    try {
+      IDataSet dataSet = getDataSet();
+      DatabaseOperation.CLEAN_INSERT.execute(connection, dataSet);
+      initSpringSecurityUser(); // must be done after data has been inserted successfully
+    } finally {
+      connection.close();
+    }
+  }
+
+  @AfterTransaction
+  public void after() throws Exception {
+    IDatabaseConnection connection = getConnection();
+    try {
+      clearSpringSecurityUser(); // reverse the initiation of the Spring Security user
+      IDataSet dataSet = getDataSet();
+      DatabaseOperation.DELETE_ALL.execute(connection, dataSet);
+    } finally {
+      connection.close();
+    }
+  }
+  
+  public void initSpringSecurityUser() throws Exception {
+    BsisUserDetails user = new BsisUserDetails(userRepository.findUserById(1l));
+    TestingAuthenticationToken auth = new TestingAuthenticationToken(user, "Credentials");
+    SecurityContextHolder.getContext().setAuthentication(auth);
+  }
+
+  public void clearSpringSecurityUser() throws Exception {
+    SecurityContextHolder.getContext().setAuthentication(null);
+  }
+
+}
